+++ helper/form_actions/00_form_action.js
@@ -24,7 +24,6 @@ FormAction.makeAbstractClass();
  */
 FormAction.startNewGroup('alchemy_form_actions');
 
-
 /**
  * Add a configuration property
  *